refactor(cmg): crée un fichier ouverture-droit
